virtio-snd: add the virtio-snd device
This adds a virtio based sound card device that provides output
and input PCM streams to a virtual guest.

In order to build the virtio-snd device the VIRTIO_SND
configuration option should be enabled in the hw/audio/Kconfig
file.

The device can be provided to a guest using the following example
command lines:

1) Using ALSA backend:
   -audiodev alsa,id=sndbe,out.try-poll=off \
   -device virtio-snd-pci

   A polling mode must be disabled because QEmu is frozen until
   playback is stopped.

2) Using PulseAudio backend:
   -audiodev pa,id=sndbe,server=/run/user/1000/pulse/native \
   -device virtio-snd-pci

Signed-off-by: Igor Skalkin <Igor.Skalkin@opensynergy.com>
Signed-off-by: Anton Yakovlev <Anton.Yakovlev@opensynergy.com>